dont mind me

+++ php/deleteclass.php
@@ -13,6 +13,12 @@ checkSessionTimeout();
 
 $id = $_GET['id'];
 
+if (!isset($_POST['token']) || $_POST['token'] !== $_SESSION['token']) {
+    header($_SERVER['SERVER_PROTOCOL'] . ' 405 Method Not Allowed');
+    exit;
+}
+
+
 
 $stmt = $conn->prepare("DELETE FROM class WHERE id = ?");
 $stmt -> bind_param("i", $id);
